Restricted Moves
Now you can move only in specified directions. NO random move is allowed.

ToDo : Castling and Pawn Promotion.

Next step : attack update.

Currently it doesn't do anything to eliminate pieces on attack which is a huge deal in case you are familiar with the game. I need to add that in.

Why am I even explaining this? It's chess. Duh!